requirements.txt
python-dotenv>=1.0
sqlalchemy>=2.0
psycopg2-binary>=2.9
requests>=2.31
websockets>=12.0
python-dateutil>=2.8
playwright>=1.40
beautifulsoup4>=4.12
lxml>=5.0
prefect>=3.0
pandas>=2.1
scipy>=1.11
statsmodels>=0.14
scikit-learn>=1.4
rapidfuzz>=3.5
streamlit>=1.30
plotly>=5.18
